requirements.txt
google-adk
requests
python-telegram-bot